fix: tighten TrackUpdate.rating to Option<u8>
Aligns with the existing TrackRepository::set_rating(rating:
Option<u8>) convention so the 0..=255 invariant is enforced at the
type level — serde rejects 256+ at the HTTP deserialization boundary
before the patch ever reaches the repository.

`u8` isn't natively bindable on Postgres (no unsigned integer types),
so the SQL bind widens to i64 via `i64::from`. The cast is a no-op
semantically because the u8 already guarantees the range.

Defers the matching DB-level CHECK (rating BETWEEN 0 AND 255) to
1.b.5b-PR-B on waveflow-server, where the `track` migration actually
lives.

+++ src-tauri/crates/core/src/repository/track.rs
@@ -52,10 +52,14 @@ pub struct TrackUpdate {
     pub track_number: Option<i64>,
     pub disc_number: Option<i64>,
     pub year: Option<i64>,
-    /// Raw POPM byte (0-255). `None` leaves the existing value alone;
-    /// the file-side tag write happens in `waveflow` (desktop) and is
-    /// not the server's concern.
-    pub rating: Option<i64>,
+    /// Raw POPM byte (0-255). Typed as `u8` so out-of-range values
+    /// can never reach the SQL bind — serde rejects `256+` at the
+    /// HTTP deserialization boundary before the repository sees the
+    /// patch. Matches the convention already established by
+    /// [`TrackRepository::set_rating`]. `None` leaves the existing
+    /// value alone; the file-side tag write happens in `waveflow`
+    /// (desktop) and is not the server's concern.
+    pub rating: Option<u8>,
 }
 
 /// "What to list" predicate for [`TrackRepository::list`].
